+++ src/performance-monitor/react.tsx
@@ -1,6 +1,6 @@
 import { FC, useEffect, useRef } from 'react'
 import PerformanceMonitor, { IOptions } from './index'
-import { useNonInitialEffect } from '../share/hooks'
+import { useOption } from '../share/hooks'
 import each from 'licia/each'
 import clone from 'licia/clone'
 
@@ -18,12 +18,8 @@ const LunaPerformanceMonitor: FC<IOptions> = (props) => {
     return () => performanceMonitor.current?.destroy()
   }, [])
 
-  each(['theme', 'color'], (key: keyof IOptions) => {
-    useNonInitialEffect(() => {
-      if (performanceMonitor.current) {
-        performanceMonitor.current.setOption(key, props[key])
-      }
-    }, [props[key]])
+  each(['theme', 'color', 'height'], (key: keyof IOptions) => {
+    useOption<PerformanceMonitor, IOptions>(performanceMonitor, key, props[key])
   })
 
   return <div ref={performanceMonitorRef} />