+++ src/performance-monitor/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0 (27 Nov 2024)
+
+* feat: chart height option
+* fix: render error if dpr is changed
+
 ## 1.0.0 (8 Oct 2024)
 
 * feat: use border instead of box-shadow